Extract notification code to `NotificationList`
Extract the notification code to a new `NotificationList` component and move the `notification` state to the game context.
